Move CLI example
